refactor: use hooks to clean code
