+++ package.json
@@ -63,10 +63,11 @@
     "@onekeyfe/cross-inpage-provider-types": "1.1.23",
     "@onekeyfe/extension-bridge-hosted": "1.1.23",
     "@onekeyfe/hd-ble-sdk": "0.2.43",
-    "@onekeyfe/hd-core": "0.2.43",
-    "@onekeyfe/hd-shared": "0.2.43",
-    "@onekeyfe/hd-transport": "0.2.43",
-    "@onekeyfe/hd-web-sdk": "0.2.43",
+    "@onekeyfe/hd-core": "file:.yalc/@onekeyfe/hd-core",
+    "@onekeyfe/hd-shared": "file:.yalc/@onekeyfe/hd-shared",
+    "@onekeyfe/hd-transport": "file:.yalc/@onekeyfe/hd-transport",
+    "@onekeyfe/hd-transport-http": "file:.yalc/@onekeyfe/hd-transport-http",
+    "@onekeyfe/hd-web-sdk": "file:.yalc/@onekeyfe/hd-web-sdk",
     "@onekeyfe/network-list": "1.6.54",
     "@onekeyfe/onekey-cross-webview": "1.1.23",
     "@starcoin/starcoin": "2.1.5",
