+++ packages/kit/src/views/Account/BulkCopyAddress/FetchAddressModal.tsx
@@ -15,12 +15,7 @@ import {
 } from '@onekeyhq/components';
 import { OneKeyErrorClassNames } from '@onekeyhq/engine/src/errors';
 import { getDefaultAccountNameInfoByImpl } from '@onekeyhq/engine/src/managers/impl';
-import type {
-  Account,
-  ImportableHDAccount,
-} from '@onekeyhq/engine/src/types/account';
-import ServiceDerivationPath from '@onekeyhq/kit-bg/src/services/ServiceDerivationPath';
-import { HardwareSDK } from '@onekeyhq/shared/src/device/hardwareInstance';
+import type { Account } from '@onekeyhq/engine/src/types/account';
 import {
   IMPL_COSMOS,
   INDEX_PLACEHOLDER,
@@ -34,6 +29,7 @@ import { deviceUtils } from '../../../utils/hardware';
 import { wait } from '../../../utils/helper';
 
 import { formatDerivationLabel } from './helper';
+import { useFetchSetRangeAddress } from './useFetchSetRangeAddress';
 
 import type { CreateAccountRoutesParams } from '../../../routes';
 import type { ModalScreenProps } from '../../../routes/types';
@@ -60,8 +56,6 @@ type IWalletAccounts = INetworkDerivationItem & {
   accountData: (Account & { index: number })[];
 };
 
-const FROM_INDEX_MAX = 2 ** 31;
-
 function getAccountIndex(template: string, path: string) {
   const templateParts = template.split(INDEX_PLACEHOLDER).filter(Boolean);
   let currentPath = path;
@@ -77,183 +71,42 @@ const FetchAddressModal: FC = () => {
   const { walletId, networkId, data, password } = route.params;
   const navigation = useNavigation<NavigationProps['navigation']>();
 
+  const {
+    progress: setRangeProgress,
+    progressText: setRangeProgressText,
+    previousAddress: setRangePreviousAddress,
+    generatedAccounts,
+  } = useFetchSetRangeAddress({
+    data,
+    walletId,
+    networkId,
+    password,
+  });
+
   const isHwWallet = walletId.startsWith('hw');
   const isHdWallet = walletId.startsWith('hd');
   const cancelFlagRef = useRef(false);
   const handleExportDataRef = useRef(false);
   const [progress, setProgress] = useState<number>(0);
-  const [generatedAccounts, setGeneratedAccounts] = useState<
-    ImportableHDAccount[]
-  >([]);
-  const temporaryAccountsRef = useRef<ImportableHDAccount[]>([]);
-  const [previousAddress, setPreviousAddress] = useState('');
+  const walletAccountsTemporaryAddressRef = useRef<string[]>([]);
 
   const [walletAccounts, setWalletAccounts] = useState<IWalletAccounts[]>([]);
 
-  const updateSetRangeAccountProgress = useCallback(
-    (result: any[], forceFinish?: boolean) => {
-      if (data.type !== 'setRange') return;
-      const { generateCount } = data;
+  const updateHwWalletsAccountProgress = useCallback(
+    (result: string[], forceFinish?: boolean) => {
+      if (data.type !== 'walletAccounts') return;
+      const { networkDerivations } = data;
+      const totalLength = networkDerivations.reduce(
+        (acc, cur) => acc + cur.accounts.length,
+        0,
+      );
       const value = forceFinish
         ? 1
-        : Math.floor((result.length / Number(generateCount)) * 100) / 100;
-      setGeneratedAccounts(result);
+        : Math.floor((result.length / Number(totalLength)) * 100) / 100;
       setProgress(value);
     },
     [data],
   );
-
-  const generateHDAccounts = useCallback(
-    async ({
-      start,
-      offset,
-      purpose,
-      template,
-    }: {
-      start: number;
-      offset: number;
-      purpose?: number;
-      template?: string;
-    }) => {
-      const accounts = await backgroundApiProxy.engine.searchHDAccounts(
-        walletId,
-        networkId,
-        password,
-        start,
-        offset,
-        purpose,
-        template,
-      );
-      return accounts;
-    },
-    [networkId, password, walletId],
-  );
-
-  // HD Wallet, fetch address for set range mode
-  useEffect(() => {
-    if (data.type !== 'setRange') return;
-    if (isHwWallet) return;
-    (async () => {
-      if (platformEnv.isNative) {
-        await wait(200);
-      }
-      const { fromIndex, generateCount, derivationOption } = data;
-      const template = derivationOption?.template;
-      const purpose = parseInt(
-        derivationOption?.category?.split("'/")?.[0] ?? '44',
-      );
-
-      const startIndex = Number(fromIndex) - 1;
-      let start = startIndex;
-      const pageSize = 5;
-      let finalLimit = Number(generateCount);
-      if (start + finalLimit > FROM_INDEX_MAX) {
-        finalLimit = FROM_INDEX_MAX - start;
-      }
-
-      const result: ImportableHDAccount[] = [];
-      let totalGenerateCount = 0;
-      while (start < startIndex + finalLimit) {
-        let offset = pageSize;
-        if (start + pageSize > startIndex + finalLimit) {
-          offset = finalLimit - totalGenerateCount;
-        }
-        if (offset <= 0) {
-          break;
-        }
-        if (cancelFlagRef.current) {
-          break;
-        }
-        const accountsWithNormalizeAddress = await generateHDAccounts({
-          start,
-          offset,
-          purpose,
-          template,
-        });
-        const accounts = await Promise.all(
-          accountsWithNormalizeAddress.map(async (account) => ({
-            ...account,
-            displayAddress: await backgroundApiProxy.engine.getDisplayAddress(
-              networkId,
-              account.displayAddress,
-            ),
-          })),
-        );
-        result.push(...accounts);
-        start += pageSize;
-        totalGenerateCount += offset;
-        if (platformEnv.isNative) {
-          await wait(100);
-        }
-        if (accounts.length !== offset) {
-          updateSetRangeAccountProgress(result, true);
-          break;
-        }
-        updateSetRangeAccountProgress(result);
-      }
-    })();
-  }, [
-    data,
-    isHwWallet,
-    networkId,
-    password,
-    walletId,
-    generateHDAccounts,
-    updateSetRangeAccountProgress,
-  ]);
-
-  // Fetch Hardware Address
-  useEffect(() => {
-    if (data.type !== 'setRange') return;
-    if (!isHwWallet) return;
-    (async () => {
-      const { fromIndex, generateCount, derivationOption } = data;
-      const template = derivationOption?.template;
-      if (!template) throw new Error('must be a template');
-      // const result = [];
-      // const hasNotExistAccount = false;
-      const indexes = [];
-      // eslint-disable-next-line no-plusplus
-      for (let i = 0; i < Number(generateCount || 0); i++) {
-        const index = Number(fromIndex) - 1 + i;
-        indexes.push(index);
-      }
-      try {
-        const addressInfos =
-          await backgroundApiProxy.serviceDerivationPath.batchGetHWAddress({
-            networkId,
-            walletId,
-            indexes,
-            template,
-            confirmOnDevice: true,
-          });
-        updateSetRangeAccountProgress(addressInfos);
-      } catch (e: any) {
-        debugLogger.common.info('getHWAddressByTemplate error: ', e);
-        const { className } = e || {};
-        if (className === OneKeyErrorClassNames.OneKeyHardwareError) {
-          deviceUtils.showErrorToast(e);
-        } else {
-          ToastManager.show({
-            title: intl.formatMessage(
-              {
-                id: 'msg__cancelled_during_the_process',
-              },
-              { type: 'error' },
-            ),
-          });
-        }
-        updateSetRangeAccountProgress(temporaryAccountsRef.current, true);
-      }
-    })();
-  }, [
-    data,
-    isHwWallet,
-    networkId,
-    walletId,
-    updateSetRangeAccountProgress,
-    intl,
-  ]);
 
   const updateWalletsAccountProgress = useCallback(
     (result: IWalletAccounts[], forceFinish?: boolean) => {
@@ -342,17 +195,6 @@ const FetchAddressModal: FC = () => {
                 account.address = hdAccount[0].displayAddress;
               }
             }
-            // if (isHwWallet) {
-            //   const address = await backgroundApiProxy.engine.getHWAddress(
-            //     account.id,
-            //     networkId,
-            //     walletId,
-            //   );
-            //   if (address !== account.address) {
-            //     throw new Error('Not same address');
-            //   }
-            //   setPreviousAddress(address);
-            // }
             const address = await backgroundApiProxy.engine.getDisplayAddress(
               networkId,
               account.address,
@@ -519,45 +361,11 @@ const FetchAddressModal: FC = () => {
     navigation,
   ]);
 
-  const isSubscribeRef = useRef(false);
-  useEffect(() => {
-    if (!isHwWallet || isSubscribeRef.current) return;
-    isSubscribeRef.current = true;
-    HardwareSDK.on('UI_EVENT', (e) => {
-      const { type, payload } = e;
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      const payloadData = payload.data as unknown as {
-        address: string;
-        path: string;
-      };
-      if (type === 'ui-previous_address_result') {
-        setPreviousAddress(payloadData.address ?? '');
-        backgroundApiProxy.serviceDerivationPath
-          .convertPlainAddressItemToImportableHDAccount({
-            networkId,
-            ...payloadData,
-          })
-          .then((account) => {
-            if (data.type === 'setRange') {
-              temporaryAccountsRef.current.push(account);
-              updateSetRangeAccountProgress(temporaryAccountsRef.current);
-            } else {
-              // updateWalletsAccountProgress(temporaryAccountsRef.current);
-            }
-          });
-      }
-    });
-  }, [
-    isHwWallet,
-    networkId,
-    updateSetRangeAccountProgress,
-    data.type,
-    // updateWalletsAccountProgress,
-  ]);
-
   useEffect(() => {
     if (handleExportDataRef.current) return;
-    if (progress >= 1) {
+    // TODO: replace
+    // if (progress >= 1) {
+    if (setRangeProgress >= 1) {
       handleExportDataRef.current = true;
       let exportData: IExportAddressData[];
       if (data.type === 'setRange') {
@@ -604,6 +412,7 @@ const FetchAddressModal: FC = () => {
       );
     }
   }, [
+    setRangeProgress,
     progress,
     data,
     generatedAccounts,
@@ -617,12 +426,13 @@ const FetchAddressModal: FC = () => {
 
   const progressText = useMemo(() => {
     if (data.type === 'setRange') {
-      const total = data.generateCount || '0';
-      return `${
-        Number.isSafeInteger(generatedAccounts.length)
-          ? generatedAccounts.length
-          : 0
-      }/${total}`;
+      // const total = data.generateCount || '0';
+      // return `${
+      //   Number.isSafeInteger(generatedAccounts.length)
+      //     ? generatedAccounts.length
+      //     : 0
+      // }/${total}`;
+      return setRangeProgressText;
     }
     if (data.type === 'walletAccounts') {
       const total = data.networkDerivations.reduce(
@@ -636,14 +446,7 @@ const FetchAddressModal: FC = () => {
       return `${finish}/${total}`;
     }
     return '';
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    data,
-    generatedAccounts.length,
-    walletAccounts,
-    walletAccounts.length,
-    progress,
-  ]);
+  }, [data, walletAccounts, setRangeProgressText]);
 
   return (
     <Modal
@@ -674,7 +477,7 @@ const FetchAddressModal: FC = () => {
       <Center w="full" h="full">
         <Box mt={6}>
           <Progress.Circle
-            progress={progress}
+            progress={setRangeProgress}
             text={
               <Center>
                 <Text typography={{ sm: 'DisplayMedium', md: 'DisplayLarge' }}>
@@ -687,7 +490,7 @@ const FetchAddressModal: FC = () => {
         <Text mt={6} typography={{ sm: 'Heading', md: 'Heading' }}>
           {intl.formatMessage({ id: 'title__fetching_addresses' })}
         </Text>
-        {previousAddress && (
+        {setRangePreviousAddress && (
           <VStack mt={4} space={1}>
             <Text
               textAlign="center"
@@ -702,7 +505,7 @@ const FetchAddressModal: FC = () => {
               typography={{ sm: 'Body2', md: 'Body2' }}
               wordBreak="break-all"
             >
-              {previousAddress}
+              {setRangePreviousAddress}
             </Text>
           </VStack>
         )}
